bid method added to service

+++ backend/src/main/kotlin/com/group7/artshare/entity/Bid.kt
@@ -15,7 +15,7 @@ class Bid {
     var id: Long = 0L
 
     @Column
-    var bidAmount: Int = 0
+    var bidAmount: Double = 0.0
 
     @ManyToOne(fetch = FetchType.EAGER, cascade = [CascadeType.ALL])
     @JoinColumn(name = "bidder")
